refactor(cmd): give each boot phase a name, and the job runner its own file
The two process roles' run() functions were 227 and 180 body lines of
linear boot sequence. Each phase now has a name and a file:
cmd/api/boot.go (bind the installation, the declared surface options,
the overlay lane, the inline relay, the model surfaces, the worker
handoff, serve-until-signal) and cmd/worker/boot.go (the debug
subcommands, deployment load, the bus, the event lanes, relay-until-
signal). startJobRunner splits into cmd/worker/jobrunner.go: the
vault/registry prologue and bounded drain stay in startJobRunner, the
one JobRunnerConfig literal becomes newJobRunner.

run 227→77 (api), 180→75 (worker), startJobRunner 127→46, and
oas30 walk 83→58.

Boot order is the contract, so it is unchanged: extensions register
before anything serves, EnsureInstallation still precedes the listener,
and every defer stays in run() — the two that had to move return their
cleanup instead (the overlay Redis close, the relay stop), so shutdown
still unwinds overlay → schema pool → pool. The JobRunnerConfig literal
stays a single call argument rather than field assignments, because
TestEveryCensusedSiteRidesALaneAProcessRoleWires reads the call: the
first cut assigned fields and quietly unwired twelve censused AI sites
from that gate's view.

Two doc comments that sat on the wrong function are reattached
(startJobRunner's was on gmailWatchConfig, envOr's was orphaned at the
end of main.go).

Groundwork for craft static --strict.
